refactor(statusline): separate rendering and status sources
